Support for multiple pages added
