+++ src/main.rs
@@ -1,7 +1,7 @@
-use back::Menu;
+use back::App;
 use iced::Application;
 use iced::Settings;
 
 fn main() -> iced::Result {
-    Menu::run(Settings::default())
+    App::run(Settings::default())
 }
